Output to stderr apart from the URL

+++ post-image.sh
@@ -4,10 +4,14 @@ cloud_bucket=${CLOUD_BUCKET:-"cloud.jamesoff.net"}
 cloud_domain=${CLOUD_DOMAIN:-"cloud.jamesoff.net"}
 cloud_queue=${CLOUD_QUEUE:-"https://sqs.eu-west-1.amazonaws.com/108685319098/cloud-output"}
 
+echoerr() {
+	echo "$@" 1>&2
+}
+
 usage() {
-	echo "usage: $0 PATH"
-	echo
-	echo "Host an image in the cloud"
+	echoerr "usage: $0 PATH"
+	echoerr ""
+	echoerr "Host an image in the cloud"
 	exit 1
 }
 
@@ -16,21 +20,21 @@ if [[ $# == 0 ]]; then
 fi
 
 if ! hash jq 2>/dev/null; then
-	echo "Needs more jq."
+	echoerr "Needs more jq."
 	exit 1
 fi
 
 file=$1
 
 if [[ ! -r $file ]]; then
-	echo "File $file does not exist or is not readable."
+	echoerr "File $file does not exist or is not readable."
 	exit 1
 fi
 
 filename=$(basename "$file")
 
 if ! aws s3 cp "$file" "s3://$cloud_bucket/assets/$filename" > /dev/null; then
-	echo "Upload failed!"
+	echoerr "Upload failed!"
 	exit 1
 fi
 
@@ -44,7 +48,7 @@ while [[ $attempts -lt 6 ]]; do
 	)
 
 	if [[ -z $message_json ]]; then
-		echo "Failed to receive message with image information. Retrying..."
+		echoerr "Failed to receive message with image information. Retrying..."
 		continue
 	fi
 
@@ -54,13 +58,19 @@ while [[ $attempts -lt 6 ]]; do
 	if [[ $body =~ ^$filename= ]]; then
 		break
 	else
-		echo "Received a message but it wasn't for us!"
+		echoerr "Received a message but it wasn't for us!"
+		body=""
 		continue
 	fi
 done
 
+if [[ -z "$body" ]]; then
+	echoerr "Failed to fetch URL."
+	exit 1
+fi
+
 cloud_path=$( echo "$body" | cut -d= -f2 )
-echo "https://$cloud_domain/$cloud_path"
+echo -n "https://$cloud_domain/$cloud_path"
 
 aws sqs delete-message \
 	--queue-url "$cloud_queue" \
